refactor: extract shared ObjectType factory logic into base/types.py
Move construct_fields, construct_self_referenced_fields, and the
create_graphene_generic_class factory into graphene_mongo/base/types.py.
The unified factory accepts executor, registry factories, and connection
field class as keyword params, eliminating ~130 lines of duplication
between the sync and async type systems.

sync/types.py and async/types.py now just call the factory with their
specific params and set get_node (sync vs async) via classmethod
attribute assignment to avoid triggering __init_subclass_with_meta__.
